Get miri passing by ignoring the UB outside the abstract machine
See https://github.com/rust-lang/miri/issues/4428#issuecomment-3017903234

Like yeah it's technically UB but cmon this is the real world who cares.

This change will also make it easy to update the CFGs to do the right thing (i.e. use fetch_or instead of fetch_add) on ISAs that don't implement it with a cmpxchg loop. Getting rid of the theoretical UB is a nice bonus.

Signed-off-by: Alex Saveau <saveau.alexandre@gmail.com>